+++ timeoff_custom/report/hr_leave_report_used.py
@@ -5,10 +5,10 @@ class HrLeaveReportUsed(models.Model):
     _auto = False
 
     name = fields.Char(string="Employee Name", readonly=True)
-    annual_leave = fields.Float(string="Annual Leave", readonly=True)
-    replacement_day_off = fields.Float(string="Replacement Day Off", readonly=True)
-    sick = fields.Float(string="Sick", readonly=True)
-    unpaid_leave = fields.Float(string="Unpaid Leave", readonly=True)
+    annual_leave = fields.Integer(string="Annual Leave", readonly=True)
+    replacement_day_off = fields.Integer(string="Replacement Day Off", readonly=True)
+    sick = fields.Integer(string="Sick", readonly=True)
+    unpaid_leave = fields.Integer(string="Unpaid Leave", readonly=True)
 
     def init(self):
         tools.drop_view_if_exists(self._cr, 'hr_leave_report_used')
@@ -18,10 +18,10 @@ class HrLeaveReportUsed(models.Model):
                 select
                     row_number() OVER () as id,
                     used.name,
-                    sum(case when used.day_off_type ~* '(?<!\w)(?:annual)(?!\w)' then used.used_days end) as annual_leave,
-                    sum(case when used.day_off_type ~* '(?<!\w)(?:replacement)(?!\w)' then used.used_days end) as replacement_day_off,
-                    sum(case when used.day_off_type ~* '(?<!\w)(?:sick)(?!\w)' then used.used_days end) as sick,
-                    sum(case when used.day_off_type ~* '(?<!\w)(?:unpaid)(?!\w)' then used.used_days end) as unpaid_leave
+                    sum(case when used.day_off_type ~* '(?<!\w)(?:annual)(?!\w)' then used.used_days end)::integer as annual_leave,
+                    sum(case when used.day_off_type ~* '(?<!\w)(?:replacement)(?!\w)' then used.used_days end)::integer as replacement_day_off,
+                    sum(case when used.day_off_type ~* '(?<!\w)(?:sick)(?!\w)' then used.used_days end)::integer as sick,
+                    sum(case when used.day_off_type ~* '(?<!\w)(?:unpaid)(?!\w)' then used.used_days end)::integer as unpaid_leave
                 from (
                 select
                     e.name,
